Move externalToolPseudonymEntityFactory to pseudonym module
