+++ apps/server/src/modules/pseudonym/controller/api-test/pseudonym.api.spec.ts
@@ -6,13 +6,13 @@ import { HttpStatus, INestApplication } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { SchoolEntity } from '@shared/domain/entity';
 import { cleanupCollections } from '@testing/cleanup-collections';
-import { externalToolPseudonymEntityFactory } from '@testing/factory/external-tool-pseudonym.factory';
 import { schoolEntityFactory } from '@testing/factory/school-entity.factory';
 import { UserAndAccountTestFactory } from '@testing/factory/user-and-account.test.factory';
 import { TestApiClient } from '@testing/test-api-client';
 import { UUID } from 'bson';
 import { Response } from 'supertest';
 import { ExternalToolPseudonymEntity } from '../../entity';
+import { externalToolPseudonymEntityFactory } from '../../testing';
 import { PseudonymResponse } from '../dto';
 
 describe('PseudonymController (API)', () => {
