+++ apps/server/src/modules/pseudonym/repo/external-tool-pseudonym.repo.integration.spec.ts
@@ -8,13 +8,13 @@ import { Page, Pseudonym } from '@shared/domain/domainobject';
 import { LegacyLogger } from '@src/core/logger';
 import { cleanupCollections } from '@testing/cleanup-collections';
 import { pseudonymFactory } from '@testing/factory/domainobject';
-import { externalToolPseudonymEntityFactory } from '@testing/factory/external-tool-pseudonym.factory';
 import { pseudonymEntityFactory } from '@testing/factory/pseudonym.factory';
 import { userFactory } from '@testing/factory/user.factory';
 import { v4 as uuidv4 } from 'uuid';
 import { PseudonymSearchQuery } from '../domain';
 import { ExternalToolPseudonymEntity } from '../entity';
 import { ExternalToolPseudonymRepo } from './external-tool-pseudonym.repo';
+import { externalToolPseudonymEntityFactory } from '../testing';
 
 describe('ExternalToolPseudonymRepo', () => {
 	let module: TestingModule;
