+++ apps/server/src/modules/pseudonym/testing/index.ts
@@ -1,0 +1,1 @@
+export * from './external-tool-pseudonym.factory';
